Refactor to use generic data service

+++ src/app/shared/facades/user.facade.ts
@@ -19,6 +19,9 @@ export class UserFacade {
   }
 
   public async delete(id: number): Promise<void> {
+    if (!id) {
+      return;
+    }
     await this.#userService.delete(id);
     await this.getAll();
   }
@@ -29,6 +32,9 @@ export class UserFacade {
   }
 
   public async getById(id: number): Promise<User | undefined> {
+    if (!id) {
+      return Promise.resolve({} as User);
+    }
     return await this.#userService.getById(id);
   }
 
@@ -37,7 +43,6 @@ export class UserFacade {
       // if not search term, return empty array.
       return Promise.resolve([]);
     }
-
     return await this.#userService.search(term);
   }
